Added new tickets fragment. Switching from activities to fragments complete.
