+++ Ticketing/.idea/misc.xml
@@ -18,6 +18,7 @@
         <entry key="..\:/Users/Chris/projects/internal-ticketing/Ticketing/app/src/main/res/layout/fragment_closed_tickets.xml" value="0.1171875" />
         <entry key="..\:/Users/Chris/projects/internal-ticketing/Ticketing/app/src/main/res/layout/fragment_counter.xml" value="0.1171875" />
         <entry key="..\:/Users/Chris/projects/internal-ticketing/Ticketing/app/src/main/res/layout/fragment_main.xml" value="0.1171875" />
+        <entry key="..\:/Users/Chris/projects/internal-ticketing/Ticketing/app/src/main/res/layout/fragment_new_ticket.xml" value="0.11354166666666667" />
         <entry key="..\:/Users/Chris/projects/internal-ticketing/Ticketing/app/src/main/res/layout/ticket.xml" value="0.1" />
         <entry key="..\:/Users/Chris/projects/internal-ticketing/Ticketing/app/src/main/res/menu/main.xml" value="0.1171875" />
         <entry key="..\:/Users/Chris/projects/internal-ticketing/Ticketing/app/src/main/res/mipmap-anydpi-v26/ic_launcher.xml" value="0.1245" />
