+++ Ticketing/app/src/main/AndroidManifest.xml
@@ -12,7 +12,6 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".NotesActivity" />
         <activity
             android:name=".MainActivity"
             android:exported="true">
