+++ Ticketing/app/src/main/res/layout/activity_main.xml
@@ -53,7 +53,7 @@
 
 
         <TextView
-            android:id="@+id/notes"
+            android:id="@+id/new_ticket"
             style="@style/bottomButtons"
             android:text="@string/new_tix"/>
